refactor(vertex-ai): Move Vertex AI client to its own package
