+++ packages/langchain_google/lib/langchain_google.dart
@@ -1,6 +1,5 @@
 /// Google module for LangChain.dart.
 library;
 
-export 'src/clients/clients.dart';
 export 'src/doc_loaders/doc_loaders.dart';
 export 'src/llms/llms.dart';
