--- packages/langchain_google/lib/src/clients/clients.dart
@@ -1,1 +1,0 @@
-export 'vertex_ai/vertex_ai.dart';
