--- packages/langchain_google/lib/src/clients/vertex_ai/vertex_ai.dart
@@ -1,2 +1,0 @@
-export 'models/models.dart';
-export 'vertex_ai_model_garden_client.dart';
