+++ packages/vertex_ai/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.0.1-dev.1
+
+- Bootstrap package.
