+++ packages/vertex_ai/example/vertex_ai_example.dart
@@ -1,0 +1,3 @@
+void main() {
+  // TODO
+}
