+++ packages/vertex_ai/lib/src/gen_ai/apis/apis.dart
@@ -1,0 +1,3 @@
+export 'text.dart';
+export 'text_chat.dart';
+export 'text_embeddings.dart';
