+++ packages/vertex_ai/lib/src/gen_ai/gen_ai.dart
@@ -1,0 +1,3 @@
+export 'apis/apis.dart';
+export 'gen_ai_client.dart';
+export 'models/models.dart';
