+++ packages/vertex_ai/lib/src/gen_ai/models/models.dart
@@ -1,4 +1,4 @@
-export 'chat.dart';
 export 'model.dart';
 export 'text.dart';
+export 'text_chat.dart';
 export 'text_embeddings.dart';
