+++ packages/vertex_ai/lib/vertex_ai.dart
@@ -1,0 +1,4 @@
+/// Vertex AI API client.
+library;
+
+export 'src/gen_ai/gen_ai.dart';
